advancedapp/app/requirements.txt
dash==2.16.1
Flask==3.0.2
pandas==2.2.1
psutil==5.9.8
mysql-connector-python==8.3.0
Flask-SQLAlchemy==3.1.1